Update generation.py
Isn't it better to first sort the probabilities and the indices together, then compute the cumulative sum of the probabilities? The mask is then created by setting all the elements in the cumulative sum that are greater than or equal to p to 1. The sorted probabilities are then set to 0 where the mask is 1. Finally, the indices where the mask is 0 are selected, and the index of the maximum value in this set is returned.